fix: server side last updated text
